Ported State and Position loggers to the new abstract logger class.
