+++ src/ch/epfl/leb/sass/simulator/generators/realtime/Fluorophore.java
@@ -20,8 +20,8 @@
 package ch.epfl.leb.sass.simulator.generators.realtime;
 
 import java.util.Random;
-import ch.epfl.leb.sass.simulator.StateLogger;
-import ch.epfl.leb.sass.simulator.PositionLogger;
+import ch.epfl.leb.sass.simulator.loggers.StateLogger;
+import ch.epfl.leb.sass.simulator.loggers.PositionLogger;
 
 /**
  * A general fluorescent molecule which emits light.
