+++ src/ch/epfl/leb/sass/simulator/loggers/StateLogger.java
@@ -16,6 +16,14 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+package ch.epfl.leb.sass.simulator.loggers;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.FileWriter;
+import java.io.PrintWriter;
+import java.util.ArrayList;
+
 /**
  * Records the fluorophore states to a file.
  *
@@ -23,40 +31,38 @@
  *
  * @author Kyle M. Douglass
  */
-
-package ch.epfl.leb.sass.simulator;
-
-import org.json.JSONObject;
-import java.io.File;
-import java.io.IOException;
-import java.io.FileWriter;
-import java.io.PrintWriter;
-import java.util.ArrayList;
-
-public class StateLogger {
+public class StateLogger extends AbstractLogger {
     private static StateLogger uniqueInstance = new StateLogger();
     
     /**
-     * The name of the log file.
-     */
-    private String filename = "";
-    
-    /**
-     * Determines whether the StateLogger is active or not.
-     */
-    private boolean performLogging = false;
-    
-    /**
-     * State information for recording.
+     * Array of emitter IDs.
      */
     private ArrayList<Integer> ids = new ArrayList();
+    
+    /**
+     * Array of times spent within a particular state.
+     */
     private ArrayList<Double> elapsedTimes = new ArrayList();
+    
+    /**
+     * Array of states that the emitter transitioned out of.
+     */
     private ArrayList<Integer> initialStates = new ArrayList();
+    
+    /**
+     * Array of states that the emitter transitioned into.
+     */
     private ArrayList<Integer> nextStates = new ArrayList();
 
+    /**
+     * Constructor is private because StateLogger is a singleton.
+     */
     private StateLogger() {
     }
 
+    /**
+     * @return An instance of the singleton.
+     */
     public static StateLogger getInstance() {
 	return uniqueInstance;
     }
@@ -86,39 +92,6 @@ public class StateLogger {
         return this.filename;
     }
     
-    /**
-     * Set the filename for logging the fluorophore state transitions and create the file.
-     * @param inFilename The full path and filename of the log file
-     * @throws IOException
-     */
-    public void setFilename(String inFilename) throws IOException {
-        // Don't do anything if input filename equals the logger's current one
-        if (inFilename.equals(this.filename))
-            return;
-        
-        // Get a unique filename if the current one exists
-        File logFile = new File(inFilename);
-        int fileId = 1;
-        String[] tokens;
-        tokens = logFile.getName().split("\\.(?=[^\\.]+$)");
-        String baseName;
-        while (logFile.exists()) {
-            // Remove last file type from filename and append a unique integer.
-            // NOTE: example.tar.gz will become example.tar_1.gz
-            // https://stackoverflow.com/questions/4545937/java-splitting-the-filename-into-a-base-and-extension#4546093
-            baseName = tokens[0] + "_" + String.valueOf(fileId);
-            
-            if (2 == tokens.length)
-                baseName += "." + tokens[1];
-            
-            baseName = logFile.getParent() + File.separator + baseName;
-            logFile = new File(baseName);
-            fileId++;
-        }
-        
-        this.filename = logFile.toString();
-    }
-    
     public ArrayList<Integer> getIds() {
         return this.ids;
     }
@@ -135,18 +108,11 @@ public class StateLogger {
         return this.nextStates;
     }
     
-    public boolean getPerformLogging() {
-        return performLogging;
-    }
-    
-    public void setPerformLogging(boolean isActive) {
-        performLogging = isActive;
-    }
-    
     /**
      * Saves the state of the logger to a file.
      * @throws IOException 
      */
+    @Override
     public void saveLogFile() throws IOException {
         File logFile = new File(this.filename);
         if ( !(logFile.exists()) )
@@ -173,6 +139,7 @@ public class StateLogger {
     /**
      * Resets the logger to its initial state.
      */
+    @Override
     public void reset() {
         this.filename = "";
         this.performLogging = false;
